In Meks DnD equipment placement moves out unhittables
